// src/app/models/patient.model.ts
export interface Patient {
  name: string;
  dob: string;
  sex: string;
  residence: string;
  mrn: string;
  idNumber: string;
  ssnLast4: string;
  phone: string;
  email: string;
  hospital: string;
  department: string;
  physician: string;
  conditions: string;
  nextAppointment: string;
  avatar: string;
}